+++ src/objects/liquid.cpp
@@ -44,7 +44,7 @@ Liquid::Liquid(FileReader& props)
   
   for(std::vector<float>::size_type i = 2; i < heightfield1->size()-2; ++i)
     {
-      (*heightfield1)[i] = sin(i/float(heightfield1->size()) * 10.0f) * 0.5f
+      (*heightfield1)[i] = sinf(i/static_cast<float>(heightfield1->size()) * 10.0f) * 0.5f
         + sinf(i / static_cast<float>(heightfield1->size()) * 5.0f) * .5f;
       (*heightfield2)[i] = (*heightfield1)[i];
     }
@@ -81,14 +81,14 @@ Liquid::draw(SceneContext& sc)
           }
 
         array->color(Color(0.5f, 0.5f, 1.0f, 0.7f));
-        array->texcoord((i * 32.0f/SAMPLES) * texscale + sin(t + i/10.0f)*0.2f,
+        array->texcoord((i * 32.0f / static_cast<float>(SAMPLES)) * texscale + sinf(t + i/10.0f)*0.2f,
                         (-32.0f * (*heightfield1)[i] + 8.0f) * texscale + sinf(t+i/10.0f)*0.2f);
-        array->vertex(i * 32.0f/SAMPLES, -32.0f * (*heightfield1)[i] + 8.0f);
+        array->vertex(i * 32.0f / static_cast<float>(SAMPLES), -32.0f * (*heightfield1)[i] + 8.0f);
 
         array->color(Color(c, c, 1.0f, 1.0f));
-        array->texcoord((i * 32.0f/SAMPLES) * texscale + sin(t + i/10.0f)*0.2f,
+        array->texcoord((i * 32.0f / static_cast<float>(SAMPLES)) * texscale + sinf(t + i/10.0f)*0.2f,
                         (-32.0f * (*heightfield1)[i]) * texscale + sinf(t+i/10.0f)*0.2f);
-        array->vertex(i * 32.0f/SAMPLES, -32.0f * (*heightfield1)[i]);
+        array->vertex(i * 32.0f / static_cast<float>(SAMPLES), -32.0f * (*heightfield1)[i]);
       }
 
     sc.color().draw(array);
@@ -103,14 +103,14 @@ Liquid::draw(SceneContext& sc)
     for(std::vector<float>::size_type i = 0; i < heightfield1->size(); ++i)
       {
         array->color(Color(0.0f, 0.0f, 0.5f, 0.7f));
-        array->texcoord((i * 32.0f/SAMPLES) * texscale + sinf(t + i/10.0f)*0.2f,
+        array->texcoord((i * 32.0f / static_cast<float>(SAMPLES)) * texscale + sinf(t + i/10.0f)*0.2f,
                         (64.0f) * texscale + sinf(t+i/10.0f)*0.2f);
-        array->vertex(i * 32.0f/SAMPLES, 64.0f);
+        array->vertex(i * 32.0f/static_cast<float>(SAMPLES), 64.0f);
 
         array->color(Color(0.5f, 0.5f, 1.0f, 0.7f));
-        array->texcoord((i * 32.0f/SAMPLES) * texscale + sin(t + i/10.0f)*0.2f,
+        array->texcoord((i * 32.0f / static_cast<float>(SAMPLES)) * texscale + sinf(t + i/10.0f) * 0.2f,
                         (-32.0f * (*heightfield1)[i] + 8.0f) * texscale + sinf(t+i/10.0f)*0.2f);
-        array->vertex(i * 32.0f/SAMPLES, -32.0f * (*heightfield1)[i] + 8.0f);
+        array->vertex(i * 32.0f/static_cast<float>(SAMPLES), -32.0f * (*heightfield1)[i] + 8.0f);
       }
 
     sc.color().draw(array);
