+++ src/screen/game_session.cpp
@@ -191,7 +191,8 @@ GameSessionImpl::draw()
   if (pause)
     {
       if ((SDL_GetTicks() / 1000) % 2)
-        Fonts::current()->vera20->draw(Vector2f(Display::get_width()/2.0f, Display::get_height()/2.0f), 
+        Fonts::current()->vera20->draw(Vector2f(static_cast<float>(Display::get_width())  / 2.0f,
+                                                static_cast<float>(Display::get_height()) / 2.0f), 
                                        "Pause");
     }
 }
