+++ src/tile/tile_packer.cpp
@@ -95,10 +95,10 @@ TilePacker::pack(const SoftwareSurface& image, int x, int y, int w, int h)
   
   assert_gl("updating tilepacker texture");
 
-  Rectf rect(Vector2f(static_cast<float>(impl->x_pos + 1) / impl->width, 
-                      static_cast<float>(impl->y_pos + 1) / impl->height), 
-             Sizef(static_cast<float>(TILE_RESOLUTION) / impl->width, 
-                   static_cast<float>(TILE_RESOLUTION) / impl->height));
+  Rectf rect(Vector2f(static_cast<float>(impl->x_pos + 1) / static_cast<float>(impl->width), 
+                      static_cast<float>(impl->y_pos + 1) / static_cast<float>(impl->height)), 
+             Sizef(static_cast<float>(TILE_RESOLUTION) / static_cast<float>(impl->width), 
+                   static_cast<float>(TILE_RESOLUTION) / static_cast<float>(impl->height)));
 
   // we move by TILE_RESOLUTION+1 to avoid tiles bleeding into each other
   // when blending
